:sparkles: (ECS) add init system
