+++ lib/ecs/systems/eat.dart
@@ -1,12 +1,10 @@
-import 'dart:math';
-
 import 'package:snake_game/ecs/components/eatable.dart';
 import 'package:snake_game/ecs/components/eater.dart';
 import 'package:snake_game/ecs/components/position.dart';
 import 'package:snake_game/ecs/components/spawnable.dart';
 import 'package:snake_game/ecs/entities/apple.dart';
 import 'package:snake_game/ecs/entities/snake.dart';
-import 'package:snake_game/ecs/systems/main.dart';
+import 'package:snake_game/ecs/systems/init.dart';
 import 'package:snake_game/ecs/systems/system.dart';
 
 class EatSystem extends System {
@@ -20,9 +18,7 @@ class EatSystem extends System {
       eatableEntities.forEach((eatable) {
         eater.hasEaten = eater.leadPosition.equal(eatable.leadPosition);
         if (eater.hasEaten && eatable is SpanwableComponent) {
-          eatable.leadPosition = Coordinates(
-              x: Random().nextInt(BOARD_SIZE - 1) + 1,
-              y: Random().nextInt(BOARD_SIZE - 1) + 1);
+          eatable.leadPosition = InitSystem.getRandomCoordinates();
         }
       });
     });
