+++ lib/widgets/game_board.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:snake_game/ecs/components/position.dart';
+import 'package:snake_game/ecs/systems/init.dart';
 import 'package:snake_game/ecs/systems/main.dart';
 
 class GameBoard extends StatelessWidget {
